+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG for 0.9.0 RC
 ======================
 
+2.6.6 (11.12.2019)
+-----
+- allow drivers to handle low level error exceptions 
+- qualify "Connection reset by peer" as connection exception 
+
 2.6.5 (11.12.2019)
 -----
 - added support for SELECT FOR UPDATE statements